Hard-fork tamayo/laravel-scout-elastic
This library isn't maintained well enough to suit our needs.
We might not be using Scout to interface with Elasticsearch
going forward, but while we are, we might as well contribute
to this library. The original project doesn't accept PRs in
a timely fashion, so hard-fork it is.

Feel free to add the following remote as `upstream` in git:

https://github.com/ErickTamayo/laravel-scout-elastic